split further
